# swagger-output.json
{
  "swagger": "2.0",
  "info": { "contact": {}, "title": "Reclamos", "version": "1.0.0" },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/api/login": {
      "post": {
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "properties": {
                "contrasenia": { "example": "123456", "type": "string" },
                "correoElectronico": {
                  "example": "admin@admin.com",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Auth"],
        "description": "Loguearse",
        "operationId": "loguearse",
        "summary": "Loguearse"
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Ingresar 'Bearer' seguido del token. Ejemplo: 'Bearer abcde12345'"
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}